Add `seq_inhibit` flag to swayrd configuration
The `seq_inhibit` flag toggles whether swayrd disables LRU updates
during window cycling command sequences. When the flag is set to `false`
(the default) windows are locked-in to the LRU as long as they remain
focused long enough. When the flag is `true` any window cycling command
is considered to start a sequence during which no LRU updates are
performed. The sequence ends upon reception of any
non-window-cycling-command (and the window with the focus at that point
in time is immediately locked-in).